Rename JwtRealmTests to JwtRealmTestCase. Move test methods to JwtRealmAuthenticateTests or JwtRealmGenerateTests. The generate smoke test realm code is now in its own class.
